+++ lib/src/great_circle_distance_base.dart
@@ -15,7 +15,15 @@ class GreatCircleDistance {
     double longitude2;
 
     GreatCircleDistance.fromRadians(
-        {this.latitude1, this.longitude1, this.latitude2, this.longitude2});
+        {this.latitude1, this.longitude1, this.latitude2, this.longitude2}) {
+        this.latitude1 = latitude1;
+        this.longitude1 = longitude1;
+
+        this.latitude2 = latitude2;
+        this.longitude2 = longitude2;
+
+        _throwExceptionOnInvalidCoordinates();
+    }
 
     GreatCircleDistance.fromDegrees(
         {this.latitude1, this.longitude1, this.latitude2, this.longitude2}) {
@@ -24,10 +32,8 @@ class GreatCircleDistance {
 
         this.latitude2 = _radiansFromDegrees(latitude2);
         this.longitude2 = _radiansFromDegrees(longitude2);
-    }
 
-    double _radiansFromDegrees(final double degrees) {
-        return degrees * (PI / 180.0);
+        _throwExceptionOnInvalidCoordinates();
     }
 
     /// Calculate distance using the Haversine formula
@@ -40,7 +46,8 @@ class GreatCircleDistance {
     /// Calculate distance using Spherical law of cosines
     /// See [Spherical law of cosines](https://en.wikipedia.org/wiki/Spherical_law_of_cosines)
     double sphericalLawOfCosinesDistance() {
-        return Spherical_LawOfCosines.distance(this.latitude1, this.longitude1, this.latitude2, this.longitude2);
+        return Spherical_LawOfCosines.distance(
+            this.latitude1, this.longitude1, this.latitude2, this.longitude2);
     }
 
     /// Calculate distance using Vincenty formula
@@ -51,4 +58,40 @@ class GreatCircleDistance {
         return Vincenty.distance(this.latitude1, this.longitude1, this.latitude2, this.longitude2);
     }
 
+    double _radiansFromDegrees(final double degrees) => degrees * (PI / 180.0);
+
+    /// A coordinate is considered invalid if it meets at least one of the following criteria:
+    ///
+    /// - Its latitude is greater than 90 degrees or less than -90 degrees.
+    ///- Its longitude is greater than 180 degrees or less than -180 degrees.
+    bool _isValidCoordinate(double latitude, longitude) =>
+        _isValidLatitude(latitude) && _isValidLongitude(longitude);
+
+    /// A latitude is considered invalid if its is greater than 90 degrees or less than -90 degrees.
+    bool _isValidLatitude(double latitudeInRadians) =>
+        !(latitudeInRadians < _radiansFromDegrees(-90.0) ||
+            latitudeInRadians > _radiansFromDegrees(90.0));
+
+    /// A longitude is considered invalid if its is greater than 180 degrees or less than -180 degrees.
+    bool _isValidLongitude(double longitudeInRadians) =>
+        !(longitudeInRadians < _radiansFromDegrees(-180.0) ||
+            longitudeInRadians > _radiansFromDegrees(180.0));
+
+    void _throwExceptionOnInvalidCoordinates() {
+        String invalidDescription = """
+            A coordinate is considered invalid if it meets at least one of the following criteria:
+            - Its latitude is greater than 90 degrees or less than -90 degrees.
+            - Its longitude is greater than 180 degrees or less than -180 degrees.
+            
+            see https://en.wikipedia.org/wiki/Decimal_degrees
+        """;
+
+        if (!_isValidCoordinate(this.latitude1, this.longitude1))
+            throw new FormatException(
+                "Invalid coordinates at latitude1|longitude1\n$invalidDescription");
+
+        if (!_isValidCoordinate(this.latitude2, this.longitude2))
+            throw new FormatException(
+                "Invalid coordinates at latitude2|longitude2\n$invalidDescription");
+    }
 }